pyproject.toml
[tool.poetry]
name = "imageedit"
version = "2021.0.2"
license = "mit"
description = "Create various icon masks and shading effects with the imageedit library"
authors = ["FredHappyface"]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Intended Audience :: Education",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Multimedia :: Graphics",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities"
]
homepage = "https://github.com/FredHappyface/Python.ImageEdit"
repository = "https://github.com/FredHappyface/Python.ImageEdit"
documentation = "https://github.com/FredHappyface/Python.ImageEdit/blob/master/README.md"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
metprint = "<2022,>=2020.6.1"
layeredimage = "<2022,>=2020.5.6"
svgtrace = "<2022,>=2020.0.2"
blendmodes = "<2022,>=2020.2.1"
colourswatch = "<2022,>=2020.1"
Pillow = "<10,>=8.1.2"

[tool.pylint.basic]
argument-naming-style = "camelCase"
attr-naming-style = "camelCase"
function-naming-style = "camelCase"
method-naming-style = "camelCase"
variable-naming-style = "camelCase"

[tool.pylint.format]
indent-string = "\t"

[tool.pylint.master]
ignore-patterns = "test_.*?py"

[tool.pylint.messages_control]
enable = ["F", "E", "W", "R", "C"]
disable = [
	"pointless-string-statement",
	"superfluous-parens",
	"bad-continuation",
	"python3"
]

[tool.black]
line-length = 100
target-version = ["py37"]

[tool.isort]
profile = "black"
indent = "Tab"

[tool.pydocstyle]
convention = "google"
ignore = "D205,D415"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
